add cmake, deal with warnings

+++ include/valuetypes.hpp
@@ -497,7 +497,7 @@ public:
     void update()
     {
         last.value = static_cast<int32_t>(value.size()) - 1;
-        count.value = value.size();
+        count.value = static_cast<int32_t>(value.size());
     }
 
 public:
